bot/requirements.txt
aiogram==3.15.0
python-dotenv
psycopg2-binary 
httpx